requirements.txt

Django==3.0
Pillow==5.2.0
djangorestframework==3.11.0
django-recaptcha2==1.4.1
django-cors-headers==3.2.1
psycopg2==2.8.4
social-auth-core==3.3.3
social-auth-app-django==3.1.0